backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
google-generativeai
groq
openai
slowapi
python-dotenv
pydantic-settings
pillow
aiofiles
email-validator
google-auth
google-auth-oauthlib
psycopg2-binary
